Refactor pagination.
