+++ app/controllers/bike_share_app.rb
@@ -1,4 +1,8 @@
+require_relative 'pagination'
+
 class BikeShareApp < Sinatra::Base
+
+  include Pagination
 
   get '/stations' do
     @stations = Station.all
@@ -125,33 +129,6 @@ class BikeShareApp < Sinatra::Base
 
 
 ############### Trip CRUD ###################
-
-  ITEMS_PER_PAGE = 30
-  
-  def on_page(items, page_num)
-    range      = (((page_num - 1) * ITEMS_PER_PAGE)..(page_num * ITEMS_PER_PAGE - 1))
-    items[range]
-  end
-
-  def total_pages(item_count)
-    (item_count / 30.to_f).ceil
-  end
-  
-  def next_page(page_num, item_count)
-    if page_num == total_pages(item_count)
-      page_num
-    else
-      page_num + 1
-    end
-  end
-
-  def previous_page(page_num)
-    if page_num == 1
-      page_num
-    else
-      page_num - 1
-    end     
-  end
 
   get '/trips' do
     redirect '/trips/page/1'
