+++ app/models/next_previous.rb
@@ -1,0 +1,10 @@
+module NextPrevious
+  def next
+    self.class.where("id > ?", id).first
+  end
+
+  def previous
+    self.class.where("id < ?", id).last
+  end
+
+end
