+++ app/models/trip.rb
@@ -1,15 +1,18 @@
-require 'time'
+require_relative 'next_previous'
 
 class Trip < ActiveRecord::Base
   belongs_to  :condition
   belongs_to  :station
   belongs_to  :subscription_type
+
   validates   :duration,             presence: true
   validates   :start_date,           presence: true
   validates   :bike_id,              presence: true
   validates   :zip_code,             presence: true
   validates   :station_id,           presence: true
   validates   :subscription_type_id, presence: true
+
+  include NextPrevious
 
   def end_station
     Station.find(end_station_id)
@@ -21,14 +24,6 @@ class Trip < ActiveRecord::Base
 
   def end_information
     "End: #{end_date}, #{end_station.name rescue "Station Not Found"}"
-  end
-
-  def next
-    Trip.where("id > ?", id).first
-  end
-
-  def previous
-    Trip.where("id < ?", id).last
   end
 
 end
